backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1
python-multipart==0.0.12
openai==1.54.4
python-dotenv==1.0.1
pydantic==2.10.1
pydantic-settings==2.6.1
aiofiles==24.1.0
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
tiktoken==0.8.0
numpy==1.26.4
